CLient-ArithmaticOperation/pyproject.toml
[project]
name = "arithmetic-localmcp"
version = "0.1.0"
description = "Arithmetic MCP server with client"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0.2",
    "langchain-core>=1.2.17",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-groq>=0.3.0",
    "python-dotenv>=1.2.2",
]